backend/requirements.txt
# FastAPI and server
fastapi==0.122.1
uvicorn[standard]==0.38.0

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.9
alembic==1.17.2

# LangChain (compatible versions - updated for compatibility)
langchain==1.1.0
langchain-core==1.1.0
langchain-openai==1.1.0
langchain-google-genai==3.2.0

# LangGraph
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10

# LangSmith
langsmith==0.4.49

# OpenAI SDK (updated for langchain-openai compatibility)
openai==2.8.1

# Authentication
python-jose[cryptography]==3.5.0
bcrypt==4.0.1  # Compatible with passlib 1.7.4 (5.0.0 causes AttributeError warning)
passlib[bcrypt]==1.7.4

# Data validation
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.1.0

# Environment variables
python-dotenv==1.2.1

# CORS
python-multipart==0.0.20

# Testing
pytest==8.4.2
pytest-asyncio==0.24.0  # Updated for Python 3.12 compatibility
pytest-cov==4.1.0
requests==2.32.5

# Web Search
tavily-python==0.7.13

# Google Gemini (using direct SDK, not LangChain wrapper)
google-generativeai==0.8.5  # Direct SDK for both text and image analysis

# Image Processing
Pillow==12.0.0
